fix the compile error at 70_dev in mac (#117)
RDC-fix the compile error of null png file, RDC-6085:
  copy the missed png file from 49_dev

RDC-6085: fix the error caused by wrong mac macro define:
  redeine the macro define according to 70_dev only support SDK 10.12+

RDC-6165: fix the error caused by a c++ function calling a oc function in .m file:
  use extern "C" to call a oc function in .cc file

RDC-6085: fix the error of lacking the implementation of class OSCrypt:
  TODO it

fix the error caused by unsupported scoped_ptr, replace scoped_ptr by unique_ptr; also fix some including file path error:
  abandon scope_ptr

fix the error of write the impletation of SaveLoginForEnterplorerImpl twice:
  delete duplicated code

fix the error of lack or misuse of micro IE_REDCORE:
  make IE_REDCORE embrace the codes which should only compiles in WINDOWS IE

fix the error of miss of micro WIN32
  modify the related gn file

fix the error of not implemented func
  TODO it

mac warning or little error:
  fix the warning in clang compiler

RDC-6242: fix the error caused by incorrectly merging code from dev_49:
  merge the incorrect code to correct code position

+++ chrome/browser/ysp_login/ysp_login_manager.cc
@@ -3686,7 +3686,7 @@ void YSPLoginManager::InitPinKeys() {
   PrefService* prefs = g_browser_process->local_state();
   const base::ListValue* pin_list = prefs->GetList(prefs::kYSPPinKeys);
   std::vector<std::string> pin_keys;
-  for (int index = 0; index < pin_list->GetSize(); index++) {
+  for (int index = 0; (size_t)index < pin_list->GetSize(); index++) {
     const base::Value* key_value = NULL;
     if (pin_list->Get(index, &key_value)) {
       pin_keys.push_back(key_value->GetString());
